# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "接口文档",
    "title": "v1接口.",
    "version": "0.0.1"
  },
  "host": "cxz.125i.cn",
  "basePath": "/v1",
  "paths": {
...
    "/order/build": {
      "post": {
        "description": "下单",
        "tags": [
          "订单"
        ],
        "operationId": "Order_Build",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "address_id": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "AddressID"
                },
                "notice": {
                  "description": "买家留言",
                  "type": "string",
                  "x-go-name": "Notice"
                },
                "product_info": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/OrderProductInfo"
                  },
                  "x-go-name": "ProductInfo"
                },
                "week_number": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "WeekNumber"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OrderBuildResp"
          }
        }
      }
    },
    "/order/detail": {
      "get": {
        "description": "订单详情",
        "tags": [
          "订单"
        ],
        "operationId": "Order_GetByOrderID",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "OrderID",
            "name": "order_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OrderGetByOrderIDResp"
          }
        }
      }
    },
    "/order/list": {
      "get": {
        "description": "订单列表",
        "tags": [
          "订单"
        ],
        "operationId": "Order_GetList",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Page",
            "name": "page",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OrderGetListResp"
          }
        }
      }
    },
    "/order/products/freight": {
      "post": {
        "description": "获取商品运费",
        "operationId": "Order_GetFreight",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "product_info": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/OrderProductInfo"
                  },
                  "x-go-name": "ProductInfo"
                },
                "week_number": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "WeekNumber"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OrderGetFreightResp"
          }
        }
      }
    },
    "/order/wxapayment": {
      "post": {
        "description": "小程序支付",
        "tags": [
          "订单"
        ],
        "operationId": "Order_WXAPayment",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "order_id": {
                  "type": "string",
                  "x-go-name": "OrderID"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OrderWXAPaymentResp"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Order": {
      "description": "订单",
      "type": "object",
      "properties": {
        "body": {
          "type": "string",
          "x-go-name": "Body"
        },
        "created": {
          "description": "创建时间",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Created"
        },
        "exp_time": {
          "description": "支付截止时间",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExpTime"
        },
        "freight": {
          "description": "运费",
          "type": "number",
          "format": "double",
          "x-go-name": "Freight"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "notice": {
          "description": "买家留言",
          "type": "string",
          "x-go-name": "Notice"
        },
        "order_address": {
          "$ref": "#/definitions/OrderAddress"
        },
        "order_id": {
          "type": "string",
          "x-go-name": "OrderID"
        },
        "order_products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderProduct"
          },
          "x-go-name": "OrderProducts"
        },
        "payment_method": {
          "description": "支付方式 1: 微信支付",
          "type": "integer",
          "format": "int64",
          "x-go-name": "PaymentMethod"
        },
        "payment_price": {
          "description": "支付金额",
          "type": "number",
          "format": "double",
          "x-go-name": "PaymentPrice"
        },
        "payment_time": {
          "description": "支付时间",
          "type": "integer",
          "format": "int64",
          "x-go-name": "PaymentTime"
        },
        "plan_id": {
          "description": "如果是月够的则有值",
          "type": "integer",
          "format": "int64",
          "x-go-name": "PlanID"
        },
        "price": {
          "description": "金额",
          "type": "number",
          "format": "double",
          "x-go-name": "Price"
        },
        "status": {
          "$ref": "#/definitions/OrderStatus"
        },
        "update_time": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UpdateTime"
        },
        "user_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "github.com/hfdend/cxz/models"
    },
    "OrderAddress": {
      "type": "object",
      "properties": {
        "address_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "AddressID"
        },
        "created": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Created"
        },
        "detail_address": {
          "type": "string",
          "x-go-name": "DetailAddress"
        },
        "district_code": {
          "type": "string",
          "x-go-name": "DistrictCode"
        },
        "district_name": {
          "type": "string",
          "x-go-name": "DistrictName"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "order_id": {
          "type": "string",
          "x-go-name": "OrderID"
        },
        "phone": {
          "type": "string",
          "x-go-name": "Phone"
        },
        "user_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "github.com/hfdend/cxz/models"
    },
    "OrderProduct": {
      "type": "object",
      "properties": {
        "created": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Created"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "image": {
          "description": "图片",
          "type": "string",
          "x-go-name": "Image"
        },
        "image_src": {
          "type": "string",
          "x-go-name": "ImageSrc"
        },
        "intro": {
          "description": "介绍",
          "type": "string",
          "x-go-name": "Intro"
        },
        "iprice": {
          "description": "此项价格",
          "type": "number",
          "format": "double",
          "x-go-name": "IPrice"
        },
        "mark": {
          "description": "备注",
          "type": "string",
          "x-go-name": "Mark"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "number": {
          "description": "商品数量",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Number"
        },
        "order_id": {
          "type": "string",
          "x-go-name": "OrderID"
        },
        "price": {
          "description": "售价",
          "type": "number",
          "format": "double",
          "x-go-name": "Price"
        },
        "product_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ProductID"
        },
        "taste": {
          "description": "味道",
          "type": "string",
          "x-go-name": "Taste"
        },
        "type": {
          "description": "类型",
          "type": "string",
          "x-go-name": "Type"
        },
        "unit": {
          "description": "商品规格",
          "type": "string",
          "x-go-name": "Unit"
        },
        "updated": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Updated"
        }
      },
      "x-go-package": "github.com/hfdend/cxz/models"
    },
    "OrderProductInfo": {
      "type": "object",
      "properties": {
        "number": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Number"
        },
        "product_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ProductID"
        }
      },
      "x-go-package": "github.com/hfdend/cxz/modules"
    },
    "OrderStatus": {
      "description": "订单状态\n1: 等待付款\n2: 付款成功\n3: 订单发货中 (月够订单)\n4: 发货完成",
      "type": "integer",
      "format": "int64",
      "x-go-package": "github.com/hfdend/cxz/models"
    },
    "Pager": {
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Count"
        },
        "limit": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Limit"
        },
        "offset": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Offset"
        },
        "page": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Page"
        },
        "page_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "PageCount"
        }
      },
      "x-go-package": "github.com/hfdend/cxz/models"
    },
    "PayWxaRequest": {
      "description": "微信小程序调用支付对象",
      "type": "object",
      "properties": {
        "nonceStr": {
          "type": "string",
          "x-go-name": "NonceStr"
        },
        "package": {
          "type": "string",
          "x-go-name": "Package"
        },
        "paySign": {
          "type": "string",
          "x-go-name": "PaySign"
        },
        "signType": {
          "type": "string",
          "x-go-name": "SignType"
        },
        "timeStamp": {
          "type": "string",
          "x-go-name": "TimeStamp"
        }
      },
      "x-go-package": "github.com/hfdend/cxz/payment/wxpay"
    },
...
  },
  "responses": {
...
    "OrderBuildResp": {
      "description": "订单详情",
      "schema": {
        "$ref": "#/definitions/Order"
      }
    },
    "OrderGetByOrderIDResp": {
      "description": "订单详情",
      "schema": {
        "$ref": "#/definitions/Order"
      }
    },
    "OrderGetFreightResp": {
      "description": "运费",
      "schema": {
        "type": "object",
        "properties": {
          "freight": {
            "description": "运费",
            "type": "number",
            "format": "double",
            "x-go-name": "Freight"
          }
        }
      }
    },
    "OrderGetListResp": {
      "description": "订单列表",
      "schema": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Order"
            },
            "x-go-name": "List"
          },
          "pager": {
            "$ref": "#/definitions/Pager"
          }
        }
      }
    },
    "OrderWXAPaymentResp": {
      "description": "小程序支付",
      "schema": {
        "$ref": "#/definitions/PayWxaRequest"
      }
    },
...
  }
}